+++ src/sklearn_raster/ufunc/__init__.py
@@ -1,0 +1,4 @@
+from ._base import FeaturewiseUfunc
+from ._meta import Dimension, Output
+
+__all__ = ["Dimension", "Output", "FeaturewiseUfunc"]